Move TrackedQueue to LibationUiBase
